Cache subclasses during rule selection
